Done till the db migrations

+++ REST API/My_First_REST_API/resources/Item.py
@@ -1,31 +1,35 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort # type: ignore
 from sqlalchemy.exc import SQLAlchemyError
 from models import ItemModel
 from db import db
 from schemas import ItemSchema, UpdateItemSchema
-
+from flask_jwt_extended import jwt_required, get_jwt
 # This is file is the same as the store.py But 
 # In this file their is just operations on the Items in the store 
 
 blp = Blueprint("Items", __name__, description="Operations on item")
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
 
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"Message": "item deleted"}
-
+    
+    @jwt_required()
     @blp.arguments(UpdateItemSchema)
     @blp.response(200, ItemSchema)
     def put(self, new_item_data, item_id):
@@ -49,11 +53,12 @@ class ItemList(MethodView):
     # so while returning before we are using {"items" : list(items.values())}
     # But now ItemSchema(many=True) at the beginning 
     # while returning already it convert it into lis so, just return 
+    @jwt_required()
     @blp.response(200,ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all() # will now return a list of items not a object with list of items
 
-    
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
